--- c0321g1_gaming_backend/src/main/java/c0321g1_gaming/controller/security/AuthRestController.java
@@ -1,4 +1,0 @@
-package c0321g1_gaming.controller.security;
-
-public class AuthRestController {
-}
