+++ c0321g1_gaming_backend/src/main/java/c0321g1_gaming/controller_service/computer/ComputerManufacturerRestController.java
@@ -1,4 +1,4 @@
-package c0321g1_gaming.controller.computer;
+package c0321g1_gaming.controller_service.computer;
 
 public class ComputerManufacturerRestController {
 }
