+++ c0321g1_gaming_backend/src/main/java/c0321g1_gaming/controller_service/computer/ComputerTypeRestController.java
@@ -1,0 +1,4 @@
+package c0321g1_gaming.controller_service.computer;
+
+public class ComputerTypeRestController {
+}
