+++ c0321g1_gaming_backend/src/main/java/c0321g1_gaming/model/repository/services/IServicesRepository.java
@@ -12,34 +12,29 @@ import javax.transaction.Transactional;
 
 @Repository
 public interface IServicesRepository extends JpaRepository<Services, Long> {
-
+    //    phap
     @Query(value = " select * from services where concat(code,`name`,prices) like ?1 and flag = 1 ", nativeQuery = true,
             countQuery = "select count(*) from services where concat(code,`name`,prices) like ?1 and flag = 1 ")
     Page<Services> pageServicesAll(String name, Pageable pageable);
 
+    //    phap
     @Query(value = "select * from services where (code like ?1 ) and ( `name` like ?2 ) and (prices like ?3 ) and flag = 1", nativeQuery = true,
             countQuery = "select count(*) from services where (code like ?1 ) and ( `name` like ?2 ) and (prices like ?3 ) and flag = 1")
     Page<Services> pageServicesCodeNamePrices(String code, String name, String prices, Pageable pageable);
 
-
-    @Query(value = "delete from services where (services_id = ?1 )", nativeQuery = true)
-    void deleteServicesId(Long id);
-
-    @Query(value = " select * from services where (services_id = ?1) and flag = 1", nativeQuery = true)
+    //    phap
+    @Query(value = " select * from services where ( services_id = ?1 ) and flag = 1", nativeQuery = true)
     Services findByIdServices(Long id);
 
-    @Transactional
-    @Modifying
-    @Query(value = "UPDATE services " +
-            "SET   `flag` = ?1  WHERE services_id = ?2 ", nativeQuery = true)
-    void updateFlag(int flagDelete, Long servicesId);
 
+    //    phap
     @Modifying
     @Transactional
     @Query(value = "insert into services (`code`, `flag`, `image`, `name`, `prices`, `quantity`, `unit_id`)" +
             "value ( ?1 , ?2 , ?3 , ?4 ,?5 ,?6, ?7)", nativeQuery = true)
     void saveServices(String code, int flag, String image, String name, double prices, int quantity, int unitId);
 
+//    phap
     @Modifying
     @Transactional
     @Query(value =" UPDATE services " +
