+++ c0321g1_gaming_backend/src/test/java/c0321g1_gaming/controller_service/ServicesRestController_pageServicesAll.java
@@ -1,6 +1,6 @@
-package c0321g1_gaming.controller;
+package c0321g1_gaming.controller_service;
 
-import c0321g1_gaming.controller.services.ServicesRestController;
+import c0321g1_gaming.controller_service.services.ServicesRestController;
 import c0321g1_gaming.model.entity.services.Services;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
@@ -9,7 +9,6 @@ import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMock
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
-import org.springframework.data.domain.Sort;
 import org.springframework.http.ResponseEntity;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
@@ -48,7 +47,7 @@ public class ServicesRestController_pageServicesAll {
         Assertions.assertEquals(200, statusCode);
         Assertions.assertEquals(1, servicesPage.getTotalElements());
         Assertions.assertEquals(1, servicesPage.getTotalPages());
-//        Assertions.assertEquals(1, servicesPage.getContent().get(0).getCode());
+        Assertions.assertEquals("SV-0002", servicesPage.getContent().get(0).getCode());
 //        Assertions.assertEquals(1, servicesPage.getContent().get(0).getName());
         //...
     }
